requirements.txt
argcomplete==3.2.1
attrs==23.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
docopt==0.6.2
et-xmlfile==1.1.0
fonttools==4.46.0
idna==3.6
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
markdown-it-py==3.0.0
matplotlib==3.8.2
mdurl==0.1.2
networkx==3.2.1
numpy==1.26.2
openpyxl==3.1.2
packaging==23.2
Pillow==10.1.0
pipreqs==0.4.13
platformdirs==4.1.0
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
PyYAML==6.0.1
referencing==0.32.0
requests==2.31.0
rich==13.7.0
rpds-py==0.16.2
shellingham==1.5.4
six==1.16.0
typer==0.9.0
typing_extensions==4.9.0
urllib3==2.1.0
userpath==1.9.1
yarg==0.1.9
